Set up CI with Azure Pipelines
For the dataset-api repo the wizard does not prompt to select an existing yaml file with the repo. It just loads azure-pipelines.yml. This is a know issue - see this thread: https://developercommunity.visualstudio.com/content/problem/345452/new-yaml-pipeline-creator-does-not-allow-existing.html. Note sure why it doesn't prompt for this particular repo. Maybe the swagger.yml confuses it. Need to investigate. [skip ci]