Remove test of main function
Due to instantiation of cosmos client being moved to global variable the test will fail unless
valid environment variables are set. This is fine until we run tests in CI; the test will fail
unless variables are setup against each build and used in azure.pipelines.yml files. This poses
several maintenance tasks such as updating the variables in devops when things change but also
if the tests do not clean up the resources in azure safelt then either resources will be lost or testdata could be left in live. Again these can be done by using a test database for cosmos client but the effort of all of the above for this test to run successfully is not of value and the test is already covered by another unit test which doesn't add this extra level of complexity.